% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz63.R
\name{l63_ens0}
\alias{l63_ens0}
\title{Simulate an initial ensemble}
\usage{
l63_ens0(K, l63_run, nsteps = 250, ...)
}
\arguments{
\item{K}{ensemble size}

\item{l63_run}{output from l63_simulate}
}
\value{
ndim x K matrix of initial ensemble
}
\description{
For each ensemble member, start with iid normal state and integrate for nsteps time steps (250)
}
\examples{
l63_run <- l63_simulate(100*1/5, 1/5, deltat=1/75, sig=3)
ens0 <- l63_ens0(20, l63_run)
lorenz_plot(l63_run$state.ts[101,], l63_run$y.ts[[101]], ens=ens0)
}
