% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz96.R
\name{l96_ens0}
\alias{l96_ens0}
\title{Simulate an initial ensemble with lorenz96}
\usage{
l96_ens0(K, l96_run, nsteps = 250, ...)
}
\arguments{
\item{K}{ensemble size}

\item{l96_run}{output from l96_simulate}
}
\value{
ndim x K matrix of initial ensemble
}
\description{
For each ensemble member, start with iid normal state and integrate for nsteps time steps (250)
}
\examples{
l96_run <- l96_simulate(0.1*100, 0.1, ndim=40, Forcing=8, deltat=1/1000)
ens0 <- l96_ens0(10, l96_run, 100)
lorenz_plot(l96_run$state.ts[101,], l96_run$y.ts[[101]], ens0)
}
