% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz96.R
\name{l96_integrate}
\alias{l96_integrate}
\title{Integrate state forward with lorenz96}
\usage{
l96_integrate(state, nsteps = 1, deltat = 1/1000, ndim = 40,
  Forcing = 8, ...)
}
\arguments{
\item{state}{initial state (can be an ensemble matrix of ndim x K)}

\item{nsteps}{how many steps ahead}

\item{deltat}{time interval for model integration}

\item{ndim}{dimension of system}

\item{Forcing}{external forcing on dynamics}

\item{...}{additional arguments passed to lorenz_observe (e.g. nobs in case obs_type=partial)}
}
\value{
state at the end of integration (either a vector or a matrix depending on initial state)
}
\description{
Time integration with lorenz96: implemented with a euler scheme in fortran90.
}
\examples{
current_state <- rnorm(40)
par(mfrow=c(4,2), mai=c(0.2,0.4,.5,.2))
for (i in 1:8){
 plot.ts(current_state, main=paste('Time=',i), xlab='', ylab='')
 current_state <- l96_integrate( current_state, 500, deltat=1/1000)
}
}
