% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz96.R
\name{l96_simulate}
\alias{l96_simulate}
\title{Simulate data from the lorenz96 model}
\usage{
l96_simulate(duration, freq, ndim = 40, Forcing = 8, sig = sqrt(0.5),
  obs_type = "odd", R_sig = sig, deltat = 1/1000, ...)
}
\arguments{
\item{duration}{total time to integrate model (duration=freq*(nsteps-1)  (nsteps = duration/freq + 1)}

\item{freq}{frequency of time integration (freq=duration/(nsteps-1))}

\item{ndim}{dimension of system}

\item{Forcing}{external forcing on dynamics}

\item{sig}{standard deviation of error measurments}

\item{obs_type}{(all: all sites, odd: every other site, 
partial: regular observations at nobs sites, one: unique observation in the middle)}

\item{R_sig}{standard deviation of error measurements for assimilation (default=sig)}

\item{deltat}{time interval for model integration}

\item{...}{additional arguments passed to lorenz_observe (e.g. nobs in case obs_type=partial)}
}
\value{
list with: state.ts matrix of nsteps x ndim
y.ts list of nsteps lists, each of them containing y, R, H, d and y.loc (sites which are observed)
f.propagate function to propagate the ensemble according to the model specification 
f.propagate <- function(state=initial state, nsteps=how many steps to propagate, ...)
ndim, duration and freq to keep track of some parameters
}
\description{
Simulate data from the lorenz96 model
}
\examples{
l96_run <- l96_simulate(0.1*100, 0.1, ndim=40, Forcing=8, deltat=1/1000)
lorenz_plot(l96_run$state.ts[101,], l96_run$y.ts[[101]])
}
