% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenzM2.R
\name{lm2_ens0}
\alias{lm2_ens0}
\title{Simulate an initial ensemble with lorenz MII}
\usage{
lm2_ens0(K, lm2_run, duration = 5, ...)
}
\arguments{
\item{K}{ensemble size}

\item{lm2_run}{output from lm2_simulate}
}
\value{
ndim x K matrix of initial ensemble
}
\description{
For each ensemble member, start with iid uniform(-1,1) state and integrate for duration
}
\examples{
lm2_run <- lm2_simulate(10 * .2, .2, sig=3)
ens0 <- lm2_ens0(10, lm2_run)
lorenz_plot(lm2_run$state.ts[11,], lm2_run$y.ts[[11]], ens0)
}
