% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenzM2.R
\name{lm2_integrate}
\alias{lm2_integrate}
\title{Integrate state forward with lorenz63}
\usage{
lm2_integrate(state, nsteps = 1, deltat = 1/200, ndim = 960, kint = 32,
  Forcing = 15)
}
\arguments{
\item{deltat}{time interval for model integration}

\item{ndim}{dimension of system}

\item{kint}{parameter K in model specification, related to number of waves}

\item{Forcing}{external forcing on dynamics}
}
\value{
state at the end of integration (either a vector or a matrix depending on initial state)
}
\description{
Time integration with lorenz MII: implemented with a euler scheme in fortran90.
}
\examples{
current_state <- runif(960, -1, 1)
par(mfrow=c(5,1), mai=c(0.2,0.4,.5,.2))
for (i in 1:5){
 plot.ts(current_state, main=paste('Time=',i), xlab='', ylab='')
 current_state <- lm2_integrate( current_state, 500, deltat=1/200)
}
}
