% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_all.R
\name{lorenz_as_df}
\alias{lorenz_as_df}
\title{Put state in long format}
\usage{
lorenz_as_df(state)
}
\arguments{
\item{state}{state vector}
}
\value{
tbl object with value and x(=site) as variables
}
\description{
Put state in long format
}
