% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_all.R
\name{lorenz_da_eval}
\alias{lorenz_da_eval}
\title{Compute diagnostics of cycling experiment}
\usage{
lorenz_da_eval(lorenz_run, model_run)
}
\arguments{
\item{lorenz_run}{reference run as returned from l**_simulate}

\item{model_run}{as returned from a cycling experiment}
}
\value{
tbl object with bias, mse, rmse, nobs, crps for background/analysis ensembles
}
\description{
compute error measures (bias, mse, rmse, crps) per time and 
per ensemble type (analysis, forecast)
}
