% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_all.R
\name{lorenz_heatmap}
\alias{lorenz_heatmap}
\title{plot time evolution of the state as a heatmap}
\usage{
lorenz_heatmap(l_run, vmean = 2.4, ...)
}
\arguments{
\item{l_run}{output of l**_simulate}

\item{vmean}{value of state around which to spread the color (e.g. 2.4=climatological mean of lorenz96)}

\item{...}{extra parameters in case (not used)}
}
\description{
x-axis=sites, y-axis=time, color=state
}
