% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_all.R
\name{lorenz_observe}
\alias{lorenz_observe}
\title{Observation operator}
\usage{
lorenz_observe(state, sig = sqrt(0.5), ndim, obs_type, R_sig = sig,
  nobs = ndim/10)
}
\arguments{
\item{state}{vector of state}

\item{sig}{standard deviation of error measurments}

\item{ndim}{dimension of system}

\item{obs_type}{(all: all sites, odd: every other site, 
partial: regular observations at nobs sites, one: unique observation in the middle)}

\item{R_sig}{standard deviation of error measurements for assimilation (default=sig)}
}
\description{
used in l**_simulate to produce observations
}
