% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_all.R
\name{lorenz_plot}
\alias{lorenz_plot}
\title{plot state as a line}
\usage{
lorenz_plot(lorenz_state, lorenz_obs, ens = NULL, tit = "", ylim,
  xlab = "site", ylab = "state", ...)
}
\arguments{
\item{lorenz_state}{vector of state}

\item{lorenz_obs}{list object with y and y.loc}

\item{ens}{ensemble matrix (can be passed as a list for multiple ensembles)}

\item{tit}{title of plot}

\item{ylim}{passed to plot()}

\item{xlab}{passed to plot()}

\item{ylab}{passed to plot()}

\item{...}{extra parameters passed to plot()}
}
\description{
plot state as a line
}
