% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_all.R
\name{lorenz_plot_eval}
\alias{lorenz_plot_eval}
\title{Plot time evolution of errors}
\usage{
lorenz_plot_eval(ens_eval)
}
\arguments{
\item{ens_eval}{data frame returned from lorenz_da_eval}
}
\description{
Plot time evolution of errors
}
